+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v2.0.2
 
+    - ADD: new action 'sleep' to wait for a number of milliseconds
     - FIX: fixed a bug that prevented the fields field as a one element list to be expanded correctly
 
 ## v2.0.1